[MS] Refactoring on TOS/logout and retry TOS if the app focus is taken
