+++ client/src/locales/en-US.json
@@ -206,10 +206,10 @@
             "privacyPolicyLink": "https://parsec.cloud/en/politique-de-confidentialite/",
             "update": {
                 "title": "New terms of service",
-                "subtitle": "We have updated our terms of service. Please read and accept them to continue.",
+                "subtitle": "We have updated our terms of service. Please read and accept them to continue. You will automatically be disconnected if you don't accept them.",
                 "tosButton": "Terms Of Service",
                 "url": "https://my.parsec.cloud/assets/pdf/Parsec-CGVU-fr.pdf",
-                "checkbox": "I have read and accept the new terms.",
+                "checkbox": "I have read and accept the terms of service.",
                 "yes": "Accept",
                 "no": "Refuse",
                 "confirmationMessage": "You have accepted the new terms of service.",
