+++ client/src/locales/fr-FR.json
@@ -205,8 +205,8 @@
             "privacyPolicy": "Politique de Confidentialité",
             "privacyPolicyLink": "https://parsec.cloud/fr/politique-de-confidentialite/",
             "update": {
-                "title": "Nouvelles conditions générales",
-                "subtitle": "Nous avons mis à jour les conditions générales d'utilisation. Veuillez les consulter avant de les accepter.",
+                "title": "Nouvelles conditions générales de vente de d'utilisation",
+                "subtitle": "Nous avons mis à jour les conditions générales d'utilisation. Veuillez les consulter et les accepter pour continuer. Vous serez automatiquement déconnecté si vous ne les acceptez pas.",
                 "tosButton": "Conditions Générales de Vente et d'Utilisation",
                 "url": "https://my.parsec.cloud/assets/pdf/Parsec-CGVU-fr.pdf",
                 "checkbox": "J'ai lu et j'accepte les conditions générales",
