+++ client/src/views/layouts/ConnectedLayout.vue
@@ -10,17 +10,22 @@
 </template>
 
 <script lang="ts" setup>
-import { needsMocks } from '@/parsec';
-import { getConnectionHandle } from '@/router';
-import { EventDistributor, EventDistributorKey } from '@/services/eventDistributor';
+import { getConnectionInfo, getTOS, needsMocks, logout as parsecLogout, acceptTOS } from '@/parsec';
+import { getConnectionHandle, navigateTo, Routes } from '@/router';
+import { EventData, EventDistributor, EventDistributorKey, Events } from '@/services/eventDistributor';
 import { FileOperationManagerKey } from '@/services/fileOperationManager';
-import { InformationManagerKey } from '@/services/informationManager';
-import { InjectionProvider, InjectionProviderKey } from '@/services/injectionProvider';
-import { IonContent, IonPage, IonRouterOutlet } from '@ionic/vue';
-import { inject, onMounted, provide, ref } from 'vue';
+import { Information, InformationLevel, InformationManagerKey, PresentationMode } from '@/services/informationManager';
+import { InjectionProvider, InjectionProviderKey, Injections } from '@/services/injectionProvider';
+import { IonContent, IonPage, IonRouterOutlet, modalController } from '@ionic/vue';
+import { inject, onMounted, provide, Ref, ref } from 'vue';
+import TOSModal from '@/views/organizations/TOSModal.vue';
+import useUploadMenu from '@/services/fileUploadMenu';
+import { MsModalResult, openSpinnerModal } from 'megashark-lib';
 
 const injectionProvider: InjectionProvider = inject(InjectionProviderKey)!;
+const injections: Ref<Injections | null> = ref(null);
 const initialized = ref(false);
+let intervalId: number | null = null;
 
 onMounted(async () => {
   const handle = getConnectionHandle();
@@ -31,11 +36,115 @@ onMounted(async () => {
   if (needsMocks() && !injectionProvider.hasInjections(handle)) {
     injectionProvider.createNewInjections(handle, new EventDistributor());
   }
-  const inj = injectionProvider.getInjections(handle);
+  injections.value = injectionProvider.getInjections(handle);
   // Provide the injections to children
-  provide(FileOperationManagerKey, inj.fileOperationManager);
-  provide(InformationManagerKey, inj.informationManager);
-  provide(EventDistributorKey, inj.eventDistributor);
+  provide(FileOperationManagerKey, injections.value.fileOperationManager);
+  provide(InformationManagerKey, injections.value.informationManager);
+  provide(EventDistributorKey, injections.value.eventDistributor);
   initialized.value = true;
+
+  injections.value.eventDistributor.registerCallback(
+    Events.TOSAcceptRequired | Events.LogoutRequested,
+    async (event: Events, _data?: EventData) => {
+      if (event === Events.LogoutRequested) {
+        await logout();
+      } else if (event === Events.TOSAcceptRequired) {
+        await tryOpeningTOSModal();
+      }
+    },
+  );
+
+  const connInfo = getConnectionInfo();
+  if (connInfo && connInfo.shouldAcceptTos) {
+    await showTOSModal();
+  }
 });
+
+async function tryOpeningTOSModal(): Promise<void> {
+  if ((await modalController.getTop()) || injections.value?.fileOperationManager.hasOperations()) {
+    if (intervalId === null) {
+      // Try again in 1 minute
+      intervalId = window.setInterval(async () => {
+        await tryOpeningTOSModal();
+      }, 60000);
+    }
+  }
+  else {
+    await showTOSModal();
+  }
+}
+
+async function logout(): Promise<void> {
+  const handle = getConnectionHandle();
+  if (!handle) {
+    console.error('Already logged out');
+    return;
+  }
+  const modal = await openSpinnerModal('HomePage.topbar.logoutWait');
+  const menuCtrls = useUploadMenu();
+  menuCtrls.hide();
+  // Cleaning the injections will automatically cancel the imports
+  await injectionProvider.clean(handle);
+  const logoutResult = await parsecLogout();
+  if (!logoutResult.ok) {
+    window.electronAPI.log('error', `Error when logging out: ${logoutResult.error}`);
+  }
+  await modal.dismiss();
+  await navigateTo(Routes.Home, { replace: true, skipHandle: true });
+}
+
+async function showTOSModal(): Promise<void> {
+  const result = await getTOS();
+
+  if (!result.ok) {
+    return;
+  }
+  if (result.value.perLocaleUrls.size === 0) {
+    window.electronAPI.log('warn', 'Received empty Terms of Service dictionary');
+    return;
+  }
+  const tosModal = await modalController.create({
+    component: TOSModal,
+    cssClass: 'modal-tos',
+    componentProps: {
+      tosLinks: result.value.perLocaleUrls,
+    },
+    canDismiss: true,
+    backdropDismiss: false,
+    showBackdrop: true,
+  });
+  await tosModal.present();
+  const { role } = await tosModal.onDidDismiss();
+  await tosModal.dismiss();
+
+  if (role === MsModalResult.Confirm) {
+    const acceptResult = await acceptTOS(result.value.updatedOn);
+    if (acceptResult.ok) {
+      const connInfo = getConnectionInfo();
+      if (connInfo) {
+        connInfo.shouldAcceptTos = false;
+      }
+      injections.value?.informationManager.present(
+        new Information({
+          message: 'CreateOrganization.acceptTOS.update.confirmationMessage',
+          level: InformationLevel.Info,
+        }),
+        PresentationMode.Toast,
+      );
+      // Early return here. If the user didn't accept the TOS or the acception fails,
+      // we will log them out.
+      return;
+    } else {
+      window.electronAPI.log('error', `Error when accepting the TOS: ${acceptResult.error}`);
+      injections.value?.informationManager.present(
+        new Information({
+          message: 'CreateOrganization.acceptTOS.update.acceptError',
+          level: InformationLevel.Info,
+        }),
+        PresentationMode.Toast,
+      );
+    }
+  }
+  await logout();
+}
 </script>
